% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{abort_status}
\alias{abort_status}
\title{Informative \code{\link[httr:stop_for_status]{stop_for_status()}} Wrapper}
\usage{
abort_status(x, task = "complete request")
}
\arguments{
\item{x}{A \code{response} object}

\item{task}{\code{character}. A message to display to the use as
"Failed to \code{task}"}
}
\value{
If request was successful, the response (invisibly). Otherwise,
raised a classed http error, as generated by
\code{\link[httr:http_condition]{http_condition()}}
}
\description{
\code{abort_status()} is just like
\code{\link[httr:stop_for_status]{stop_for_status()}}, except it prints the
response message returned by the server on error (if any).
}
\keyword{internal}
