% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_bool}
\alias{assert_bool}
\title{Assert that an object is either \code{TRUE} or \code{FALSE}}
\usage{
assert_bool(x)
}
\arguments{
\item{x}{The object to check}
}
\value{
\code{x} (invisibly), or an error if failed
}
\description{
Assert that an object is either \code{TRUE} or \code{FALSE}
}
\keyword{internal}
