% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{auth}
\alias{auth}
\title{Use \code{\link[httr:authenticate]{authenticate()}} with \code{secret} Credentials}
\usage{
auth(creds = es_creds_get(), type = "basic")
}
\arguments{
\item{creds}{A \code{secret} containing a \code{user} and \code{password}}

\item{type}{type of HTTP authentication.  Should be one of the following
types supported by Curl: basic, digest, digest_ie, gssnegotiate,
ntlm, any.  It defaults to "basic", the most common type.}
}
\value{
A \code{request} object with authentication
}
\description{
Use \code{\link[httr:authenticate]{authenticate()}} with \code{secret} Credentials
}
\keyword{internal}
