% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_bool}
\alias{check_bool}
\alias{assert_bool}
\title{Check/Assert that an Object is Either \code{TRUE} or \code{FALSE}}
\usage{
check_bool(x)

assert_bool(x)
}
\arguments{
\item{x}{The object to check}
}
\value{
\verb{check_*()} returns \code{TRUE} if successful or a error message string
otherwise; \verb{assert_*()} returns \code{x} (invisibly) if successful or an
error otherwise
}
\description{
Check/Assert that an Object is Either \code{TRUE} or \code{FALSE}
}
\keyword{internal}
