% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-check_essence.R
\name{check_datasource}
\alias{check_datasource}
\alias{assert_datasource}
\title{Check/Assert that an Input is a Valid ESSENCE Datasource String}
\usage{
check_datasource(datasource, creds = es_creds_get(), fmatch = TRUE)

assert_datasource(
  datasource,
  creds = es_creds_get(),
  fmatch = TRUE,
  add = NULL
)
}
\arguments{
\item{creds}{\code{secret}. User login credentials for ESSENCE.}

\item{fmatch}{[\code{logical(1)}]\cr
Use the set operations implemented in \code{\link[fastmatch]{fmatch}} in package \pkg{fastmatch}.
If \pkg{fastmatch} is not installed, this silently falls back to \code{\link[base]{match}}.
\code{\link[fastmatch]{fmatch}} modifies \code{y} by reference:
A hash table is added as attribute which is used in subsequent calls.}

\item{add}{[\code{AssertCollection}]\cr
Collection to store assertion messages. See \code{\link[checkmate]{AssertCollection}}.}
}
\value{
\verb{check_*()} returns \code{TRUE} if successful or a error message string
otherwise; \verb{assert_*()} returns \code{x} (invisibly) if successful or an
error otherwise
}
\description{
Check/Assert that an Input is a Valid ESSENCE Datasource String
}
\keyword{internal}
