% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_ts.R
\name{check_ts_query}
\alias{check_ts_query}
\title{Check TimeSeries Query Arguments}
\usage{
check_ts_query(
  datasource,
  startDate,
  endDate,
  timeResolution,
  dots,
  creds = es_creds_get()
)
}
\arguments{
\item{datasource}{\verb{[character(1)]} An ESSENCE data source; see
\code{\link[essence:es_datasource]{es_datasource()}} for available options}

\item{startDate}{\verb{[Date(1)]} or an object that can be coerced to one.
The start date for the query.}

\item{endDate}{\verb{[Date(1)]} or an object that can be coerced to one.
The end date for the query.}

\item{timeResolution}{\verb{[character(1)]}. The resolution of the returned
timeseries.}

\item{dots}{Additional query parameters; see the
\href{https://essence.tn.gov/tn_state/usersguide/UsersGuide.jsp}{ESSENCE User Guide}
(in particular the API section) for details. All arguments must be named.}

\item{creds}{\code{secret}. User login credentials for ESSENCE.}
}
\value{
\code{TRUE} (invisibly) if assertions are passed
}
\description{
Check TimeSeries Query Arguments
}
\keyword{internal}
