% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-check_essence.R
\name{check_value}
\alias{check_value}
\alias{assert_value}
\title{Check/Assert that an Object is a Valid ESSENCE Field Value}
\usage{
check_value(
  value,
  field,
  datasource,
  creds = es_creds_get(),
  check_datasource = TRUE,
  check_field = TRUE,
  fmatch = TRUE
)

assert_value(
  value,
  field,
  datasource,
  creds = es_creds_get(),
  fmatch = TRUE,
  check_field = TRUE,
  check_datasource = TRUE,
  add = NULL
)
}
\arguments{
\item{field}{\code{character}. A field id within the given \code{datasource} (see
\code{\link[essence:es_fields]{es_fields()}} for choices)}

\item{datasource}{\verb{[character(1)]}. A datasource name
(see \code{\link[essence:es_datasources]{es_datasources()}} for choices)}

\item{creds}{\code{secret}. User login credentials for ESSENCE.}

\item{fmatch}{[\code{logical(1)}]\cr
Use the set operations implemented in \code{\link[fastmatch]{fmatch}} in package \pkg{fastmatch}.
If \pkg{fastmatch} is not installed, this silently falls back to \code{\link[base]{match}}.
\code{\link[fastmatch]{fmatch}} modifies \code{y} by reference:
A hash table is added as attribute which is used in subsequent calls.}

\item{add}{[\code{AssertCollection}]\cr
Collection to store assertion messages. See \code{\link[checkmate]{AssertCollection}}.}
}
\value{
\verb{check_*()} returns \code{TRUE} if successful or a error message string
otherwise; \verb{assert_*()} returns \code{x} (invisibly) if successful or an
error otherwise
}
\description{
Check/Assert that an Object is a Valid ESSENCE Field Value
}
\keyword{internal}
