% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creds.R
\name{es_creds}
\alias{es_creds}
\alias{es_creds_set}
\alias{es_creds_get}
\title{Manage ESSENCE Credentials}
\usage{
es_creds_set(user)

es_creds_get()
}
\arguments{
\item{user}{\code{character}. The username for your ESSENCE account.
If \code{NULL}, \code{es_creds_get()} will return the username (if set).}
}
\value{
\code{es_creds_set()} returns \code{NULL}. \code{es_creds_get()} returns the
currently stored username and password as a \code{secret}
}
\description{
\code{es_creds_set()} and \code{es_creds_get()} set and retrieve a username/password
for ESSENCE. They are wrappers around
\code{\link[keyring:key_set]{key_set()}} and
\code{\link[keyring:key_get]{key_get()}}.
}
