% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creds.R
\name{es_creds_set}
\alias{es_creds_set}
\title{Manage ESSENCE Credentials}
\usage{
es_creds_set(username)
}
\arguments{
\item{username}{\code{character}. The username for your ESSENCE account.
If \code{NULL}, \code{es_creds_get()} will return the username (if set).}
}
\value{
\code{es_creds_get()} returns your username when \code{username = NULL} and
password otherwise. \code{es_creds_set()} returns \code{NULL}.
}
\description{
\code{es_creds_set()} and \code{es_creds_get()} set and retrieve a username/password
for ESSENCE. They are wrappers around
\code{\link[keyring:key_set]{key_set()}} and
\code{\link[keyring:key_get]{key_get()}}.
}
