% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_datasources.R
\name{es_datasources}
\alias{es_datasources}
\title{Get ESSENCE Data Sources Available to User}
\usage{
es_datasources(creds = es_creds_get())
}
\arguments{
\item{creds}{\code{secret}. User login credentials for ESSENCE.}
}
\value{
A \code{tibble} with columns \code{label} (\code{chr}) and \code{src} (\code{chr}). \code{label}
contains the human-friendly label of the datasource; \code{name} contains the
name used by the API.
}
\description{
Get ESSENCE Data Sources Available to User
}
