% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_fields.R
\name{es_fields}
\alias{es_fields}
\title{Get Field Information for an ESSENCE Data Source}
\usage{
es_fields(datasource, creds = es_creds_get())
}
\arguments{
\item{datasource}{\code{character}. A datasource name
(see \code{\link[essence:es_datasources]{es_datasources()}} for choices)}

\item{creds}{\code{secret}. ESSENCE login credentials}
}
\value{
A \code{tibble} with columns \code{parentParamName}, \code{description},
\code{dataBeanId}, \code{paramName}, \code{id}, \code{label}, \code{parentValue}, \code{type}
(all \code{chr})
}
\description{
Get Field Information for an ESSENCE Data Source
}
