% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_GET.R
\name{es_GET}
\alias{es_GET}
\title{\code{\link[httr:GET]{GET()}} Shortcut for ESSENCE Queries}
\usage{
es_GET(query, creds = es_creds_get())
}
\arguments{
\item{query}{\code{string}. Query portion of the ESSENCE URL;
appended to \code{\link{es_uri}}}

\item{creds}{\code{secret}. ESSENCE login credentials.}
}
\value{
An {httr} \code{response}
}
\description{
\code{\link[httr:GET]{GET()}} Shortcut for ESSENCE Queries
}
\keyword{internal}
