% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_ts.R
\name{es_ts}
\alias{es_ts}
\title{Request TimeSeries Data from ESSENCE}
\usage{
es_ts(
  datasource,
  start,
  end = Sys.Date(),
  resolution = c("daily", "weekly", "monthly", "quarterly", "yearly"),
  ...,
  creds = es_creds_get(),
  check = TRUE
)
}
\arguments{
\item{datasource}{\verb{[character(1)]} An ESSENCE data source; see
\code{\link[essence:es_datasource]{es_datasource()}} for available options}

\item{start}{\verb{[Date(1)]} or an object that can be coerced to one. The start
date for the query.}

\item{end}{\verb{[Date(1)]} or an object that can be coerced to one. The end date
for the query.}

\item{resolution}{\verb{[character(1)]}. The resolution of the returned timeseries}

\item{...}{Additional query parameters; see the
\href{https://essence.tn.gov/tn_state/usersguide/UsersGuide.jsp}{ESSENCE User Guide}
(in particular the API section) for details. All arguments must be named.}

\item{creds}{\code{secret}. User login credentials for ESSENCE.}

\item{check}{\verb{[logical(1)]} Should the query parameters be checked before
requesting? Note that some timeseries-specific parameters (e.g. \code{detector})
will currently cause an error.}
}
\value{
A \code{tibble} with columns \code{date}, \code{count}, \code{expected}, \code{levels},
\code{colorID}, \code{color}, \code{altText}, & \code{details}
}
\description{
\code{es_ts()} retrieves the results of a timeseries API query. The provided
parameters are required for all queries; additional parameters may be
required for specific data sources. These will be listed on failure if
not provided.
}
