% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_ts.R
\name{es_ts_query}
\alias{es_ts_query}
\title{Create an ESSENCE TimeSeries Query}
\usage{
es_ts_query(
  datasource,
  startDate,
  endDate,
  timeResolution,
  ...,
  creds = es_creds_get(),
  check = TRUE
)
}
\arguments{
\item{datasource}{\verb{[character(1)]} An ESSENCE data source; see
\code{\link[essence:es_datasource]{es_datasource()}} for available options}

\item{startDate}{\verb{[Date(1)]} or an object that can be coerced to one.
The start date for the query.}

\item{endDate}{\verb{[Date(1)]} or an object that can be coerced to one.
The end date for the query.}

\item{timeResolution}{\verb{[character(1)]}. The resolution of the returned
timeseries.}

\item{...}{Additional query parameters; see the
\href{https://essence.tn.gov/tn_state/usersguide/UsersGuide.jsp}{ESSENCE User Guide}
(in particular the API section) for details. All arguments must be named.}

\item{creds}{\code{secret}. User login credentials for ESSENCE.}

\item{check}{\verb{[logical(1)]} Should the query parameters be checked before
requesting? Note that some timeseries-specific parameters (e.g. \code{detector})
will currently cause an error.}
}
\value{
A string containing the URL query parameters to pass to
\code{\link[essence:es_GET]{es_GET()}}
}
\description{
Create an ESSENCE TimeSeries Query
}
\keyword{internal}
