% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_uri.R
\docType{data}
\name{es_uri}
\alias{es_uri}
\title{ESSENCE URI Stem}
\format{
A string
}
\source{
\href{https://essence.tn.gov/tn_state/usersguide/api/documentation.jsp}{ESSENCE API Docs}
}
\usage{
es_uri
}
\description{
ESSENCE URI Stem
}
\keyword{datasets}
