% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_values.R
\name{es_values}
\alias{es_values}
\title{Get Possible Values for an ESSENCE Field}
\usage{
es_values(field, datasource, creds = es_creds_get(), check = TRUE)
}
\arguments{
\item{field}{\code{character}. A field id within the given \code{datasource} (see
\code{\link[essence:es_fields]{es_fields()}} for choices)}

\item{datasource}{\verb{[character(1)]}. A datasource name
(see \code{\link[essence:es_datasources]{es_datasources()}} for choices)}

\item{creds}{\code{secret}. User login credentials for ESSENCE.}

\item{check}{\code{logical}. Should arguments be checked before querying
ESSENCE?}
}
\value{
A \code{tibble} with columns \code{display} and \code{value} (both \code{chr})
}
\description{
Get Possible Values for an ESSENCE Field
}
