% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/essence-package.R
\docType{package}
\name{essence-package}
\alias{essence}
\alias{essence-package}
\title{essence: Interface to TN ESSENCE}
\description{
{essence} provides an R interface to select data from the TN
    ESSENCE platform.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jesse-smith/essence}
  \item Report bugs at \url{https://github.com/jesse-smith/essence/issues}
}

}
\author{
\strong{Maintainer}: Jesse Smith \email{jesse.smith@shelbycountytn.com}

}
\keyword{internal}
