% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secret.R
\name{secret-methods}
\alias{secret-methods}
\alias{format.essence_secret}
\alias{obj_str_footer.essence_secret}
\alias{names.essence_secret}
\alias{names<-.essence_secret}
\alias{vec_ptype_abbr.essence_secret}
\alias{vec_ptype_full.essence_secret}
\alias{vec_ptype2.essence_secret.essence_secret}
\alias{vec_ptype2.essence_secret.character}
\alias{vec_ptype2.character.essence_secret}
\alias{vec_cast.essence_secret.nbscar_secret}
\alias{vec_cast.essence_secret.character}
\alias{vec_cast.character.essence_secret}
\alias{$.essence_secret}
\alias{[[.essence_secret}
\title{S3 Methods for \code{secret} Objects}
\usage{
\method{format}{essence_secret}(x, ...)

\method{obj_str_footer}{essence_secret}(x, ...)

\method{names}{essence_secret}(x)

\method{names}{essence_secret}(x) <- value

\method{vec_ptype_abbr}{essence_secret}(x, ...)

\method{vec_ptype_full}{essence_secret}(x, ...)

\method{vec_ptype2}{essence_secret.essence_secret}(x, y, ...)

\method{vec_ptype2}{essence_secret.character}(x, y, ...)

\method{vec_ptype2}{character.essence_secret}(x, y, ...)

\method{vec_cast}{essence_secret.nbscar_secret}(x, to, ...)

\method{vec_cast}{essence_secret.character}(x, to, ...)

\method{vec_cast}{character.essence_secret}(x, to, ...)

\method{$}{essence_secret}(x, i)

\method{[[}{essence_secret}(x, i)
}
\description{
S3 Methods for \code{secret} Objects
}
\keyword{internal}
